+++ src/compiler/json_output.c
@@ -80,6 +80,7 @@ static inline const char *decl_type_to_string(Decl *type)
 		case DECL_CT_ASSERT: return "$assert";
 		case DECL_CT_ECHO: return "$echo";
 		case DECL_CT_EXEC: return "$exec";
+		case DECL_CT_EXPAND: return "$expand";
 		case DECL_CT_INCLUDE: return "$include";
 		case DECL_ALIAS: return "alias";
 		case DECL_TYPEDEF: return "typedef";